+++ src/rules/element-permitted-occurrences.ts
@@ -1,5 +1,5 @@
 import { Rule } from '../rule';
-import { DOMNode } from '../dom';
+import { HtmlElement } from '../dom';
 import { DOMReadyEvent } from '../event';
 import { Validator } from '../meta';
 
@@ -7,7 +7,7 @@ class ElementPermittedOccurrences extends Rule {
 	setup(){
 		this.on('dom:ready', (event: DOMReadyEvent) => {
 			const doc = event.document;
-			doc.visitDepthFirst((node: DOMNode) => {
+			doc.visitDepthFirst((node: HtmlElement) => {
 				const parent = node.parent;
 
 				if (!parent.meta){