+++ src/rules/element-permitted-order.ts
@@ -1,5 +1,5 @@
 import { Rule } from '../rule';
-import { DOMNode } from '../dom';
+import { HtmlElement } from '../dom';
 import { DOMReadyEvent } from '../event';
 import { Validator } from '../meta';
 
@@ -7,13 +7,13 @@ class ElementPermittedOrder extends Rule {
 	setup(){
 		this.on('dom:ready', (event: DOMReadyEvent) => {
 			const doc = event.document;
-			doc.visitDepthFirst((node: DOMNode) => {
+			doc.visitDepthFirst((node: HtmlElement) => {
 				if (!node.meta){
 					return;
 				}
 
 				const rules = node.meta.permittedOrder;
-				Validator.validateOrder(node.children, rules, (child: DOMNode, prev: DOMNode) => {
+				Validator.validateOrder(node.children, rules, (child: HtmlElement, prev: HtmlElement) => {
 					this.report(child, `Element <${child.tagName}> must be used before <${prev.tagName}> in this context`);
 				});
 			});